backend: split DataManagerInterface
